Abstracted fs tasks from main class.
